Added permissions model so users can only fetch their own data.

+++ quanda/api/views.py
@@ -4,6 +4,7 @@ from rest_framework import viewsets, filters
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework import status
+from guardian.shortcuts import assign_perm, get_objects_for_user
 from django.contrib.postgres.search import SearchVector
 from django.shortcuts import get_object_or_404
 from django_filters.rest_framework import DjangoFilterBackend
@@ -30,20 +31,12 @@ class TagViewSet(viewsets.ModelViewSet):
             return super().paginate_queryset(queryset)
         return None
 
+    def get_queryset(self):
+        return get_objects_for_user(self.request.user, 'view_tag', klass=Tag)
 
-class ThoughtViewSet(viewsets.ModelViewSet):
-    queryset = Thought.objects.all()
-    serializer_class = ThoughtSerializer
-    filter_backends = [filters.SearchFilter, DjangoFilterBackend]
-    search_fields = ['primary']
-    pagination_class = CustomPageNumberPagination
-    permission_classes = [IsAuthenticated]
-
-
-    def paginate_queryset(self, queryset):
-        if self.action == 'list':
-            return super().paginate_queryset(queryset)
-        return None
+    def perform_create(self, serializer):
+        tag = serializer.save()
+        assign_perm('view_tag', self.request.user, tag)
 
 
 class EntryViewSet(viewsets.ModelViewSet):
@@ -53,7 +46,7 @@ class EntryViewSet(viewsets.ModelViewSet):
     pagination_class = CustomPageNumberPagination
 
     def get_queryset(self):
-        queryset = super().get_queryset()
+        queryset = get_objects_for_user(self.request.user, 'view_entry', klass=Entry)
         search_query = self.request.query_params.get('search', None)
         if search_query:
             queryset = queryset.annotate(
@@ -63,17 +56,26 @@ class EntryViewSet(viewsets.ModelViewSet):
 
     def paginate_queryset(self, queryset):
         if self.action == 'list':
+            print(queryset.values())
             return super().paginate_queryset(queryset)
         return None
-    
+
+    def perform_create(self, serializer):
+        entry = serializer.save()
+        assign_perm('view_entry', self.request.user, entry)
+
     @action(detail=True, methods=['post'], url_path='add-thought')
     def add_thought(self, request, pk=None):
         entry = self.get_object()
+
+        if not request.user.has_perm('view_entry', entry):
+            return Response(status=status.HTTP_403_FORBIDDEN)
+
         thought_data = request.data
         thought_data['entry'] = entry.id
         serializer = ThoughtSerializer(data=thought_data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save()           
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -81,17 +83,25 @@ class EntryViewSet(viewsets.ModelViewSet):
     @action(detail=True, methods=['put'], url_path='edit-thought/(?P<thought_id>[^/.]+)')
     def edit_thought(self, request, pk=None, thought_id=None):
         entry = self.get_object()
+
+        if not request.user.has_perm('view_entry', entry):
+            return Response(status=status.HTTP_403_FORBIDDEN)
+        
         thought = get_object_or_404(Thought, id=thought_id, entry=entry)
         request.data['entry'] = entry.id
         serializer = ThoughtSerializer(thought, data=request.data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save()          
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     @action(detail=True, methods=['delete'], url_path='delete-thought/(?P<thought_id>[^/.]+)')
     def delete_thought(self, request, pk=None, thought_id=None):
         entry = self.get_object()
+
+        if not request.user.has_perm('view_entry', entry):
+            return Response(status=status.HTTP_403_FORBIDDEN)
+        
         thought = get_object_or_404(Thought, id=thought_id, entry=entry)
         thought.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
